test(desktop): consolidate issue 424 coverage
Refs #424
