+++ apps/desktop/e2e/app-launch.contract.spec.ts
@@ -55,37 +55,4 @@ test.describe('App Launch', () => {
     expect(methods.hasRemove).toBe(true);
   });
 
-  test('should expose agent IPC methods', async () => {
-    const methods = await page.evaluate(() => {
-      const sero = (window as any).sero;
-      return {
-        hasOpen: typeof sero?.agent?.open === 'function',
-        hasPrompt: typeof sero?.agent?.prompt === 'function',
-        hasAbort: typeof sero?.agent?.abort === 'function',
-        hasClose: typeof sero?.agent?.close === 'function',
-      };
-    });
-    expect(methods.hasOpen).toBe(true);
-    expect(methods.hasPrompt).toBe(true);
-    expect(methods.hasAbort).toBe(true);
-    expect(methods.hasClose).toBe(true);
-  });
-
-  test('should expose VCS IPC methods', async () => {
-    const methods = await page.evaluate(() => {
-      const sero = (window as any).sero;
-      return {
-        hasListCheckpoints: typeof sero?.vcs?.listCheckpoints === 'function',
-        hasGetState: typeof sero?.vcs?.getState === 'function',
-        hasCreateCheckpoint: typeof sero?.vcs?.createCheckpoint === 'function',
-        hasRestore: typeof sero?.vcs?.restore === 'function',
-        hasDiff: typeof sero?.vcs?.diff === 'function',
-      };
-    });
-    expect(methods.hasListCheckpoints).toBe(true);
-    expect(methods.hasGetState).toBe(true);
-    expect(methods.hasCreateCheckpoint).toBe(true);
-    expect(methods.hasRestore).toBe(true);
-    expect(methods.hasDiff).toBe(true);
-  });
 });
