--- apps/desktop/src/components/apps/explorer/explorer-language-routing.test.ts
@@ -1,20 +1,0 @@
-import { describe, expect, it } from 'vitest';
-import { getMonacoLanguageIdFromPath } from '@/lsp/language-routing';
-import { getLanguage } from './editor/editor-panel-shared';
-
-const CASES = [
-  '/workspace/src/App.tsx',
-  '/workspace/src/index.jsx',
-  '/workspace/src/main.py',
-  '/workspace/config/settings.yaml',
-  '/workspace/assets/logo.SVG',
-  '/workspace/README',
-];
-
-describe('explorer language routing', () => {
-  it.each(CASES)('keeps editor inference aligned with the canonical routing for %s', (path) => {
-    const canonicalLanguage = getMonacoLanguageIdFromPath(path);
-
-    expect(getLanguage(path)).toBe(canonicalLanguage);
-  });
-});
